Revert "Move iters.h to src"
parallel.h was apparently specifically public - undo hiding of iter.h

This reverts commit ce3e45a9a6af42393fbb4b220e9bb2c8ebabd901.
