+++ include/manifold/CMakeLists.txt
@@ -1,5 +1,6 @@
 set(MANIFOLD_PUBLIC_HDRS
 	common.h
+	iters.h
 	manifold.h
 	optional_assert.h
 	polygon.h
